requirements.txt
aniso8601==9.0.1
click==8.0.1
confluent-kafka==1.7.0
Flask==2.0.1
Flask-RESTful==0.3.9
Flask-Cors==3.0.10
flask-restx==0.5.1
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
marshmallow==3.13.0
pymongo==3.12.0
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
six==1.16.0
Werkzeug==2.0.1
